requirements.txt
django-filter
markdown
djangorestframework
Django